# A More Flexible And Lightweight CA