--- a/DataFormats/TrackSoA/src/classes.h
+++ b/DataFormats/TrackSoA/src/classes.h
@@ -5,6 +5,4 @@
 #include "DataFormats/TrackSoA/interface/TracksSoA.h"
 #include "DataFormats/TrackSoA/interface/TracksHost.h"
 
-using namespace pixelTopology;
-
 #endif  // DataFormats_TrackSoA_src_classes_h