--- a/DataFormats/TrackingRecHitSoA/src/classes.cc
+++ b/DataFormats/TrackingRecHitSoA/src/classes.cc
@@ -1,7 +1,5 @@
 #include "DataFormats/Portable/interface/PortableHostCollectionReadRules.h"
-#include "DataFormats/TrackingRecHitSoA/interface/TrackingRecHitsSoA.h"
-#include "Geometry/CommonTopologies/interface/SimplePixelTopology.h"
+#include "DataFormats/TrackingRecHitSoA/interface/TrackingRecHitsHost.h"
 
-SET_PORTABLEHOSTCOLLECTION_READ_RULES(PortableHostCollection<TrackingRecHitLayout<pixelTopology::Phase1>>);
-SET_PORTABLEHOSTCOLLECTION_READ_RULES(PortableHostCollection<TrackingRecHitLayout<pixelTopology::Phase2>>);
-SET_PORTABLEHOSTCOLLECTION_READ_RULES(PortableHostCollection<TrackingRecHitLayout<pixelTopology::HIonPhase1>>);
+using namespace reco;
+SET_PORTABLEHOSTMULTICOLLECTION_READ_RULES(HitPortableCollectionHost);