… DataFormats/TrackSoA/interface/TracksSoA.h
… DataFormats/TrackSoA/interface/alpaka/TrackUtilities.h
… DataFormats/TrackSoA/src/alpaka/classes_cuda_def.xml
… DataFormats/TrackSoA/src/alpaka/classes_rocm_def.xml
… DataFormats/TrackSoA/src/classes_def.xml
… DataFormats/TrackSoA/test/BuildFile.xml
… DataFormats/TrackSoA/test/TestReadHostTrackSoA.cc
… DataFormats/TrackSoA/test/TestWriteHostTrackSoA.cc
… DataFormats/TrackSoA/test/alpaka/TrackSoAHeterogeneous_test.cc
… DataFormats/TrackSoA/test/alpaka/TrackSoAHeterogeneous_test.dev.cc
… DataFormats/TrackSoA/test/alpaka/TrackSoAHeterogeneous_test.h
… DataFormats/TrackSoA/test/alpaka/TrajectoryStateSoA_t.cc
… DataFormats/TrackSoA/test/alpaka/TrajectoryStateSoA_t.dev.cc
… DataFormats/TrackSoA/test/alpaka/TrajectoryStateSoA_t.h
… DataFormats/TrackSoA/test/testReadHostTrackSoA.py
… DataFormats/TrackSoA/test/testWriteAndReadTrackSoA.sh
… DataFormats/TrackSoA/test/testWriteHostTrackSoA.py
… DataFormats/TrackingRecHitSoA/interface/TrackingRecHitsDevice.h
… DataFormats/TrackingRecHitSoA/interface/TrackingRecHitsHost.h
… DataFormats/TrackingRecHitSoA/interface/TrackingRecHitsSoA.h
… DataFormats/TrackingRecHitSoA/interface/alpaka/TrackingRecHitsSoACollection.h
… DataFormats/TrackingRecHitSoA/src/alpaka/classes_cuda_def.xml
… DataFormats/TrackingRecHitSoA/src/alpaka/classes_rocm_def.xml
… DataFormats/TrackingRecHitSoA/src/classes.h
… DataFormats/TrackingRecHitSoA/src/classes_def.xml
… DataFormats/TrackingRecHitSoA/test/BuildFile.xml
… DataFormats/TrackingRecHitSoA/test/TestReadHostHitSoA.cc
… DataFormats/TrackingRecHitSoA/test/TestWriteHostHitSoA.cc
… DataFormats/TrackingRecHitSoA/test/alpaka/Hits_test.cc
… DataFormats/TrackingRecHitSoA/test/alpaka/Hits_test.dev.cc